refactory 1
