+++ s.bat
@@ -1,0 +1,1 @@
+ng s -o --host 0.0.0.0
