+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { NavComponent } from './components/nav/nav.component';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-root',
@@ -7,47 +7,18 @@ import { NavComponent } from './components/nav/nav.component';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-  title = 'Julio A. Marines CRUD Alumnos y cursos';
-  public menuSelec :string = '';
-  public blank:boolean=true;
-  public display:boolean=false;
-  public crud:boolean=false;
-  public logout:boolean=false;
-  public loginOk:boolean=false;
-  public sidenav:boolean=false;
+  title = 'Julio A. Marines Administracion Alumnos y cursos';
   
-public seleccion(itemMenu:string){
-  
-  
-  switch(itemMenu){
-    case "CRUD":
-      this.display= true;
-      this.crud= false;
-      this.logout=false;
-      break;
-    case "Listado":
-        this.crud= true;
-        this.display=false;
-        this.logout=false;
-        break;
-    case "Logout":
-        this.logout= false;
-        this.display=false;
-        this.crud= false;
-        this.loginOk=false;
-        break;   
-    }
-  
-  }  
+  loginOk:boolean = false;
 
-public logOK(){
-    this.loginOk=true;
-  }
+  constructor(private router: Router){}
+  public logOK(){
+     this.loginOk=true;
+     this.router.navigate(['/entorno']);}
 
-public sideSwitch(){
-  alert('otro cambio')
-  this.sidenav=!this.sidenav;
-}  
-
+  // public logOut(){
+  //    this.loginOk=false;
+  //    this.router.navigate(['/']);
+  //  }
 
 }
