+++ src/app/app.module.ts
@@ -1,32 +1,27 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { ReactiveFormsModule } from '@angular/forms';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { NavComponent } from './components/nav/nav.component';
-import { AsideComponent } from './components/aside/aside.component';
-import { MainComponent } from './components/main/main.component';
-import { AsidenComponent } from './components/asiden/asiden.component';
-
-import { FooterComponent } from './components/footer/footer.component';
-import { LoginFormComponent } from './components/login/login-form/login-form.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { MaterialModule } from './modules/material/material.module';
+
+import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
+import { AsideComponent } from './components/aside/aside.component';
+import { AsidenComponent } from './components/asiden/asiden.component';
 import { FormAlumnosModule } from './modules/alumnos/formAlumnos/FormAlumnosModule';
 import { AlumnosModule } from './modules/alumnos/alumnos.module';
+import { EntornoModule } from './modules/entorno/entorno.module';
+import { SeguridadModule } from './modules/seguridad/seguridad.module';
+import { LoginFormComponent } from './modules/seguridad/login-form/login-form.component';
+
 @NgModule({
   declarations: [
     AppComponent,
-    NavComponent,
     AsideComponent,
-    MainComponent,
-    FooterComponent,
-    LoginFormComponent,
-    AsidenComponent
-  
-    
+    AsidenComponent,
+    LoginFormComponent
+
   ],
   imports: [
     BrowserModule,
@@ -35,7 +30,12 @@ import { AlumnosModule } from './modules/alumnos/alumnos.module';
     BrowserAnimationsModule,
     MaterialModule,
     FormAlumnosModule,
-    AlumnosModule],
+    AlumnosModule,
+    EntornoModule,
+    SeguridadModule
+  ],
+    
+    
   providers: [],
   bootstrap: [AppComponent]
 })
