--- src/app/components/main/main.component.html
@@ -1,12 +1,0 @@
-<div><h1> main </h1></div>
-<!-- < componetisar esta seccion 
-<section class="overflow-hidden text-gray-700">
-    <div class="container px-5 py-2 mx-auto lg:pt-24 lg:px-32">
-      <div class="flex flex-wrap -m-1 md:-m-2">
-        <div class="w-1/2 p-1 md:p-2"> -->
-          <!-- <img class="block object-cover object-center w-full h-full rounded-lg"
-            src="http://3.bp.blogspot.com/-9WxjByAUXb8/UR4QlX6s6hI/AAAAAAAABvs/CeB5RwocyiI/s200/martinez.jpg"> -->
-        <!-- </div>
-      </div>
-    </div>
-  </section>> -->
