+++ src/app/components/nav/nav.component.html
@@ -30,7 +30,7 @@
         <button mat-icon-button class="example-icon favorite-icon" aria-label="Example icon-button with heart icon">
           <mat-icon>favorite</mat-icon>
         </button> -->
-        <button mat-icon-button class="example-icon" aria-label="Example icon-button with share icon">
+        <button (click)="logOut()" mat-icon-button class="example-icon" aria-label="Example icon-button with share icon">
           <mat-icon>logout</mat-icon>
         </button>
       </div>
