+++ src/app/components/nav/nav.component.ts
@@ -1,5 +1,6 @@
 import { Component, EventEmitter, OnInit,Output } from '@angular/core';
 import { Router } from '@angular/router';
+
 @Component({
   selector: 'app-nav',
   templateUrl: './nav.component.html',
@@ -10,25 +11,20 @@ import { Router } from '@angular/router';
 
 export class NavComponent implements OnInit {
 
-  @Output() menuSeleccionado = new EventEmitter<string>();  
-  @Output() sideSwitchNavBarIut = new EventEmitter<string>();  
+    @Output() sideSwitchNavBarIut = new EventEmitter<string>();  
   
-
- /*  public itemMenu:string[]=["Alumnos","Cursos","Logout"] */
-
-  constructor() { }
+  constructor(private router:Router) { }
 
   ngOnInit(): void { }
 
   public sideSwitch(){
     this.sideSwitchNavBarIut.emit();
-    //this.router.navigate(['/dashboard']);
-    
+     
   }
 
-  // public seleccion(menu:string){
-  //   this.menuSeleccionado.emit(menu);
-  // }
+  public logOut(){
+        this.router.navigate(['/']);   
+  }
 }
