+++ src/app/modules/entorno/entorno.module.ts
@@ -2,12 +2,20 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { EntornoComponent } from './entorno.component';
+import { FooterComponent } from 'src/app/components/footer/footer.component';
+import { NavComponent } from 'src/app/components/nav/nav.component';
+import { MaterialModule } from '../material/material.module';
 
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
+    MaterialModule
   ],
-  declarations: [EntornoComponent]
+  declarations: [
+    EntornoComponent,
+    NavComponent,
+    FooterComponent
+  ]
 })
 export class EntornoModule { }
