+++ src/app/modules/seguridad/seguridad.module.ts
@@ -1,12 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { MaterialModule } from '../material/material.module';
+
+//import { LoginFormComponent } from './login-form/login-form.component';
+
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+   // LoginFormComponent
+  ],
   imports: [
-    CommonModule
+    CommonModule,MaterialModule
   ]
 })
 export class SeguridadModule { }
